Move untit tests to proper file
