+++ bach/tests/unit/bach/test_df_from_pandas.py
@@ -1,11 +1,14 @@
 """
 Copyright 2022 Objectiv B.V.
 """
+from datetime import datetime
+
 import pytest
 
+from bach import DataFrame, SeriesFloat64, SeriesInt64, SeriesTimestamp, SeriesString, SeriesBoolean
 from bach.from_pandas import _assert_column_names_valid
 from tests.unit.bach.test_utils import ColNameValid
-from tests.unit.bach.util import get_pandas_df
+from tests.unit.bach.util import get_pandas_df, FakeEngine
 
 
 def test__assert_column_names_valid_generic(dialect):
@@ -54,3 +57,39 @@ def test__assert_column_names_valid_db_specific(dialect):
             msg = f'Column name ".*" is not valid for SQL dialect {dialect.name}, and cannot be escaped.'
             with pytest.raises(ValueError, match=msg):
                 _assert_column_names_valid(dialect=dialect, df=pdf)
+
+
+def test_string_as_index(dialect):
+    TEST_DATA_SUPPORTED_TYPES = [
+        ['fierljeppen', 1.32, 4, datetime.datetime(2015, 12, 13, 9, 54, 45, 543), True]
+    ]
+
+    engine = FakeEngine(dialect=dialect)
+    df = DataFrame.from_pandas(
+        engine=engine,
+        df=get_pandas_df(TEST_DATA_SUPPORTED_TYPES, ['string', 'float', 'int', 'timestamp', 'bool']),
+        convert_objects=True,
+    )
+    assert isinstance(df['float'], SeriesFloat64)
+    assert isinstance(df['int'], SeriesInt64)
+    assert isinstance(df['timestamp'], SeriesTimestamp)
+    assert isinstance(df['string'], SeriesString)
+    assert isinstance(df['bool'], SeriesBoolean)
+    assert df.index_dtypes == {'_index_string': 'string'}
+
+
+def test_load_df_without_conversion(dialect):
+    TEST_DATA_SUPPORTED_TYPES = [
+        [1.32, 4, datetime.datetime(2015, 12, 13, 9, 54, 45, 543), True]
+    ]
+
+    engine = FakeEngine(dialect=dialect)
+    df = DataFrame.from_pandas(
+        engine=engine,
+        df=get_pandas_df(TEST_DATA_SUPPORTED_TYPES, ['float', 'int', 'timestamp', 'bool']),
+        convert_objects=False,
+    )
+    assert isinstance(df['float'], SeriesFloat64)
+    assert isinstance(df['int'], SeriesInt64)
+    assert isinstance(df['timestamp'], SeriesTimestamp)
+    assert isinstance(df['bool'], SeriesBoolean)
